requirements.txt
pandas
python-dotenv
sqlite3
openpyxl